process/requirements.txt
faststream[rabbit]==0.6.4
pydantic==2.12.5
aiohttp==3.13.2
aiofiles==25.1.0
python-dotenv==1.0.1
